backend/requirements.txt
absl-py==2.3.1
attrs==25.3.0
backports.functools-lru-cache==2.0.0
beautifulsoup4==4.13.5
blinker==1.9.0
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2025.8.3
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
Eel==0.18.2
Flask==3.1.2
flask-cors==6.0.1
Flask-MySQLdb==2.0.0
flatbuffers==25.2.10
fonttools==4.59.1
future==1.0.0
gevent==25.5.1
gevent-websocket==0.10.1
googletrans==4.0.0rc1
greenlet==3.2.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib_resources==6.5.2
indic-transliteration==2.3.75
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.2
Js2Py==0.74
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.5
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.3
MouseInfo==0.1.3
mysqlclient==2.2.7
nltk==3.9.1
numpy==2.2.6
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pipwin==0.5.2
playsound==1.2.2
protobuf==4.25.3
psutil==7.0.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20240210
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.19.2
pyjsparser==2.7.1
PyMsgBox==1.0.9
pynput==1.8.1
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywin32==311
regex==2025.9.1
requests==2.32.5
rfc3986==1.5.0
rich==14.1.0
roman==5.1
scikit-learn==1.7.2
scipy==1.15.3
screen_brightness_control==0.24.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SpeechRecognition==3.14.3
textblob==0.19.0
threadpoolctl==3.6.0
toml==0.10.2
tqdm==4.67.1
typer==0.19.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
wikipedia==1.4.0
WMI==1.5.1
zope.event==5.1.1
zope.interface==7.2
Flask
Flask-Cors
mysql-connector-python
bcrypt
